Decided to make changes
The health and mana system scripts have been replaced to match the one you made so there is no cause for confusion when we each go into it ourselves, a lot simpler than what I did. Attempted to add the integer called MPDrain for the special moves, though right now, it does not work. I can confirm that the HP and MP bar drains when the HP and MP values go down. I tested this by haviung the enemy damage the player is within its range. Thats all I have for now. I wont be active tonight due to my usual Thursday thing so we will talk tomorrow, I will ask Steven to help. My plan for tomorrow is to possibly add EXP and a leveling system at a basic level, like increasing attack, HP, MP and the like, which should be simple enough to implimet in the UI and in the player script.

See you tommorow